requirements.txt
torchvision==0.9.1
matplotlib==3.5.1
tensorboard==2.8.0
scipy==1.8.0
opencv-python==4.5.5.64
tqdm==4.46.0
opt_einsum==3.3.0
imageio==2.17.0
scikit-image==0.19.2
